requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
tk>=0.1.0
customtkinter>=4.6.0
geopandas>=0.10.0
folium>=0.12.0
pillow>=9.0.0
ttkthemes>=3.2.0
pyinstaller>=5.0.0
